Relocate code to fix type dependencies.
